\name{dat1m}
\alias{dat1m}
\docType{data}

\title{
Serialized glmnet model file 
}

\description{
A character string for a serialized glmnet model
}

\usage{data("dat1m")}

\format{A character string}

\details{Example glmnet model based on dat1}

\source{
The source of this function is in the
}

\references{
There are no references
}

\examples{
data(dat1m)
## maybe str(dat1m) ; plot(dat1m) ...
}

\keyword{datasetsm}
